requirements.txt
Flask==2.1.3
Werkzeug==2.0.3
gspread==5.4.0
oauth2client==4.1.3
python-dotenv==0.20.0
pytz==2023.3
requests==2.28.1
Jinja2==3.1.2
google-api-python-client==2.126.0